Add move command wti mv (#150)
* wti mv work in progress

* Not necessary

* Remove stray spaces

* Update releases notes, copyright dates and company name.

* Cleaning up

* Update image

* Fix resource

* Remove deprecated specifications in gemspec

* Reword error message

* Fix a display bugs on wti rm command

* Fixes to mv command

+++ lib/web_translate_it/command_line.rb
@@ -18,6 +18,8 @@ module WebTranslateIt
           message = "Creating master files"
         when 'rm'
           message = "Deleting files"
+        when 'mv'
+          message = "Moving files"
         when 'addlocale'
           message = "Adding locale"
         when 'rmlocale'
@@ -127,26 +129,77 @@ module WebTranslateIt
       STDOUT.sync = true
       if parameters == []
         puts StringUtil.failure("Error: You must provide the path to the master file to remove.")
-        puts "Usage: wti path/to/rm master_file_1 path/to/master_file_2 ..."
+        puts "Usage: wti rm path/to/master_file_1 path/to/master_file_2 ..."
         exit
       end
       WebTranslateIt::Connection.new(configuration.api_key) do |http|
         parameters.each do |param|
           if Util.ask_yes_no("Are you sure you want to delete the master file #{param}?\nThis will also delete its target files and translations.", false)
-            configuration.files.find_all{ |file| file.file_path == param }.each do |master_file|
-              master_file.delete(http)
-              # delete files
-              success = File.delete(master_file.file_path) if File.exists?(master_file.file_path)
-              complete_success = false if !success
-              configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
-                success = File.delete(target_file.file_path) if File.exists?(target_file.file_path)
+            files = configuration.files.find_all{ |file| file.file_path == param }
+            if files.any?
+              files.each do |master_file|
+                master_file.delete(http)
+                # delete files
+                if File.exists?(master_file.file_path)
+                  success = File.delete(master_file.file_path)
+                  puts StringUtil.success("Deleted master file #{master_file.file_path}.") if success
+                end
                 complete_success = false if !success
+                configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
+                  if File.exists?(target_file.file_path)
+                    success = File.delete(target_file.file_path)
+                    puts StringUtil.success("Deleted target file #{target_file.file_path}.") if success
+                  else
+                    puts StringUtil.failure("Target file #{target_file.file_path} doesn’t exist locally")
+                  end
+                  complete_success = false if !success
+                end
               end
+              puts StringUtil.success("All done.") if complete_success
+            else
+              puts StringUtil.failure("#{param}: File doesn’t exist on project.")
             end
           end
         end
       end
-      puts StringUtil.success("Master file deleted.")
+      complete_success
+    end
+    
+    def mv
+      complete_success = true
+      STDOUT.sync = true
+      if parameters.count != 2
+        puts StringUtil.failure("Error: You must provide the source path and destination path of the master file to move.")
+        puts "Usage: wti mv path/to/master_file_old_path path/to/master_file_new_path ..."
+        exit
+      end
+      source = parameters[0]
+      destination = parameters[1]
+      WebTranslateIt::Connection.new(configuration.api_key) do |http|
+        if Util.ask_yes_no("Are you sure you want to move the master file #{source} and its target files?", true)
+          configuration.files.find_all{ |file| file.file_path == source }.each do |master_file|
+            master_file.upload(http, false, false, nil, false, false, true, true, destination)
+            # move master file
+            if File.exists?(source)
+              success = File.rename(source, destination) if File.exists?(source)
+              puts StringUtil.success("Moved master file #{master_file.file_path}.") if success
+            end
+            complete_success = false if !success
+            configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
+              if File.exists?(target_file.file_path)
+                success = File.delete(target_file.file_path)
+                complete_success = false if !success
+              end
+            end
+            configuration.reload
+            configuration.files.find_all{ |file| file.master_id == master_file.id }.each do |target_file|
+              success = target_file.fetch(http)
+              complete_success = false if !success
+            end
+            puts StringUtil.success("All done.") if complete_success
+          end
+        end
+      end
       complete_success
     end
